Move Doc struct to separate file
