+++ include/iocore/cache/CacheVC.h
@@ -30,6 +30,7 @@
 #include "../../../src/iocore/cache/P_CacheVol.h"
 */
 #include "../../../src/iocore/cache/P_CacheHttp.h"
+#include "iocore/cache/Doc.h"
 
 // aio
 #include "iocore/aio/AIO.h"
