+++ src/iocore/cache/CacheEvacuateDocVC.cc
@@ -28,6 +28,7 @@
 
 // inkcache
 #include "iocore/cache/CacheDefs.h"
+#include "iocore/cache/Doc.h"
 #include "P_CacheHttp.h"
 #include "P_CacheInternal.h"
 #include "P_CacheVol.h"
