+++ src/iocore/cache/Stripe.cc
@@ -22,6 +22,7 @@
  */
 
 #include "iocore/cache/Cache.h"
+#include "iocore/cache/Doc.h"
 #include "P_CacheDisk.h"
 #include "P_CacheInternal.h"
 #include "P_CacheVol.h"
